ed25519 support (#10)

+++ expat/xmlparse.c
@@ -7315,7 +7315,7 @@ accountingReportDiff(XML_Parser rootParser,
 
   fprintf(stderr,
           " (+" EXPAT_FMT_PTRDIFF_T("6") " bytes %s|%d, xmlparse.c:%d) %*s\"",
-          (int)bytesMore, (account == XML_ACCOUNT_DIRECT) ? "DIR" : "EXP",
+          (ptrdiff_t)bytesMore, (account == XML_ACCOUNT_DIRECT) ? "DIR" : "EXP",
           levelsAwayFromRootParser, source_line, 10, "");
 
   { /*WATCOMC*/
